+++ changelog.md
@@ -1,11 +1,12 @@
 # MuditaOS changelog
 
-## Current release 
+## [Current release]
 
 ## Added
 
 * `[calendar]` Added a new field for date of the event when the user adds/edits event.
 * `[file indexer db]` Added messages for File Indexer db.
+* `[settings]` Added Torch window (front-end only).
 
 ## Changed